Change getBalanceFactor return type from byte to int
Balance is not stored as metadata and therefore does not benefit
from a smaller byte field. Returning it as int is more suitable
for arithmetic.

+++ src/org/sosy_lab/common/collect/PathCopyingPersistentAvlTreeMap.java
@@ -91,8 +91,8 @@ public final class PathCopyingPersistentAvlTreeMap<
       return n == null ? -1 : n.height;
     }
 
-    static byte getBalanceFactor(Node<?, ?> n) {
-      return (byte) (getHeight(n.left) - getHeight(n.right));
+    static int getBalanceFactor(Node<?, ?> n) {
+      return (getHeight(n.left) - getHeight(n.right));
     }
 
     // Methods used for path copying
